refactor: replace react-reflex
