--- packages/components/src/molecules/AllotmentPanels/index.ts
@@ -1,3 +1,0 @@
-export {default as AllotmentResizableColumnsPanel} from './AllotmentResizableColumnsPanel';
-export {default as AllotmentResizableRowsPanel} from './AllotmentResizableRowsPanel';
-export type {AllotmentResizableColumnsPanelType, AllotmentResizableRowsPanelType} from './types';
