+++ packages/components/src/molecules/Panels/ResizableRowsPanel.stories.tsx
@@ -1,13 +1,17 @@
-import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { AllRowsArgs, WithoutBottomRowArgs } from "./args";
-import ResizableRowsPanel from "./ResizableRowsPanel";
+import {ComponentMeta, ComponentStory} from '@storybook/react';
+import {AllRowsArgs, WithoutBottomRowArgs} from './args';
+import ResizableRowsPanel from './ResizableRowsPanel';
 
 export default {
-  title: "Molecules/ResizableRowsPanel",
+  title: 'Molecules/ResizableRowsPanel',
   component: ResizableRowsPanel,
 } as ComponentMeta<typeof ResizableRowsPanel>;
 
-const Template: ComponentStory<typeof ResizableRowsPanel> = (args) => <ResizableRowsPanel {...args} />;
+const Template: ComponentStory<typeof ResizableRowsPanel> = args => (
+  <div style={{height: 600, width: 1000}}>
+    <ResizableRowsPanel {...args} />
+  </div>
+);
 
 export const AllPanels = Template.bind({});
 AllPanels.args = AllRowsArgs;
