+++ packages/components/src/molecules/Panels/index.ts
@@ -1,8 +1,3 @@
-export { default as ResizableColumnsPanel } from "./ResizableColumnsPanel";
-export { default as ResizableRowsPanel } from "./ResizableRowsPanel";
-export type {
-  ResizableColumnsPanelLayout,
-  ResizableColumnsPanelType,
-  ResizableRowsPanelLayout,
-  ResizableRowsPanelType,
-} from "./types";
+export {default as ResizableColumnsPanel} from './ResizableColumnsPanel';
+export {default as ResizableRowsPanel} from './ResizableRowsPanel';
+export type {ResizableColumnsPanelType, ResizableRowsPanelType} from './types';
