+++ packages/components/src/molecules/index.ts
@@ -1,6 +1,5 @@
 export * from './ActivityBar';
 export * from './ActionsMenu';
-export * from './AllotmentPanels';
 export * from './LearnCard';
 export * from './LearnPage';
 export * from './Panels';
